Ensure consistent usage of the RN_DEBUG_STRING_CONVERTIBLE flag in C++ (#38765)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/38765

## Changelog:
[Internal] -

The `RN_DEBUG_STRING_CONVERTIBLE` was interchangeably used via both `#if` and `#ifdef` (the latter being incorrect), which led to both inconsistency, or even the code plain not compiling it `RN_DEBUG_STRING_CONVERTIBLE=0` explicitly.

Furthermore there was no guarantee it's defined, which can generate warnings with `-Wundef`/`-Wall` (or even errors with `-Werror` on top of that).

This fixes all the usages to be uniform and makes sure the flag is defined when it's used to prevent compiler warnings.

Reviewed By: christophpurrer

Differential Revision: D48021565

fbshipit-source-id: 653b6af96bb0361e6f1d62c1b545ec01309760e8