Downscale and move to a prebuilt-image analyze code/pr (#38819)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/38819

The jobs for `analyze_pr` and `analyze_code` are currently running on Android's Docker container on a XL instance. This is costly. As those jobs are relatively small, not on the critical path, and need the Android container only for `java`, I've moved the to the node-broswers CI executor:
https://circleci.com/developer/images/image/cimg/node

Changelog:
[Internal] [Changed] - Downscale and move to a prebuilt-image analyze code/pr

Reviewed By: cipolleschi

Differential Revision: D48116159

fbshipit-source-id: a12782b810fb225ab6916a35054a891e73f3a717