Add basic active pointer tracking to the pointer event processor (#38865)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/38865

Changelog: [Internal] - Add basic active pointer tracking to the pointer event processor

This diff lays the groundwork for keeping track of all the "active" pointers. The ActivePointer struct right now only contains a copy of the PointerEvent data but in future diffs this will be extended to include additional stateful metadata.

It's important to note that the code as written in this diff only "tracks" pointers when they are down and won't track pointers that are only hovering. I do want to include this in future diffs but I couldn't include that handling here as it requires changes to the iOS native pointer handling to stop deriving its own hover events which ends up snowballing even more changes. I will be making this change in (hopefully) the next diff where I refactor the hover event derivation to the fabric C++ layer.

Reviewed By: rozele

Differential Revision: D48167438

fbshipit-source-id: d1c7146e7b8d46b2f8defd4785618ee1cd54f155