Use type aliases for native module getConstants where applicable
Summary:
## Changelog:
[Internal] -

This replaces object literal return types in `getConstants()` for core RN native modules with type aliases.

This is a valid change from the point of view of binary compatibility between native/JS, and provides a couple of benefits, such as clearer code in general, but also ability to employ C++ codegen for the corresponding types and their marshalling.

bypass-github-export-checks

Reviewed By: christophpurrer

Differential Revision: D48685992

fbshipit-source-id: f21358b12448d30a7b5e25e81f62ef77964d1fcd