Initial sync of debugger-frontend (#39226)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/39226

Changelog: [Internal]

Syncs a custom build of the Chrome DevTools frontend (performed using the script in D48680624) into the `react-native/debugger-frontend` package. **The package is not used anywhere yet.** In a subsequent diff, it will be integrated into `react-native/dev-middleware`.

Review the `BUILD_INFO` file for information about the source revision and build settings that were used to generate this diff.

Reviewed By: huntie

Differential Revision: D48682302

fbshipit-source-id: 50ea532f45d8c751c141d5f4253afe080af31559